+++ RELEASE-NOTES.md
@@ -13,6 +13,7 @@
 * #76 Reduce client logging noise
 * #77 RequestTimer to log response headers at DEBUG level
 * #78 Jakarta Client - set custom ConnectorProvider via the extender
+* #79 Jakarta Client - support for Jetty HttpClient
 
 ## 3.0.M1